requirements.txt
# This file collects all required third-party applications that are needed
# to run this project. Later you can install all these apps in a row
# using pip. Example::
#
#     pip install -r requirements.pip

django==1.4.2
south==0.7.6
#django-fiber==0.9.8.9
git+https://github.com/Piemonster/django-generate-scaffold.git
#git+https://github.com/Piemonster/django-staticfiles-bootstrap.git
django-debug-toolbar
django-allauth
djrill
django-reversion>=1.6.4
django-celery
python-memcached

django-jsonfield

geopy
django-easy-maps

git+https://github.com/danfairs/django-lazysignup.git
mock
django-facebook

django-crispy-forms

git+https://github.com/Thinktiv/django-easy-friends.git
git+https://github.com/yourcelf/django-notification#egg=django-notification
httplib2
git+https://github.com/simplegeo/python-oauth2.git#egg=python-oauth2
git+https://github.com/eldarion/django-oauth-access.git#egg=django-oauth-access
gdata
git+https://github.com/pythonforfacebook/facebook-sdk.git#egg=facebook-sdk
python-twitter

django-compressor
django-mptt
pillow
git+https://github.com/lincolnloop/django-jsonit.git#egg=jsonit
git+https://github.com/lincolnloop/django-launchpad.git#egg=launchpad
django-jenkins
coverage 
pylint
git+https://github.com/twaddington/django-gravatar.git

djangorestframework
markdown  # Markdown support for the browseable API.
django-filter  # Filtering support

django-autocomplete-light